IntoChained trait returns Chain type, not Link
All methods will now take a function/closure directly, instead of simply
returning a Link type, which then had to be chained with a function,
requiring an additional method call.

This comes at a cost of using more generics in the code. But it makes
chaining a bit more ergonomic in already existing method chains,
which makes it a worthy trade-off.